Created common JS file. For each one page(home, registration) created own js file. Some function to hide sidebar and navbar on click outside of their area.
