+++ src/main/resources/static/js/registration.js
@@ -1,0 +1,12 @@
+var currentLang = '';
+var bottom = '';
+
+$(document).ready(function () {
+    if (currentLang == '') {
+        currentLang = $('#dd_menu_link').data('currentLang');
+    }
+    getLanguage();
+    setLocaleFields();
+    buildPageByCurrentLang();
+});
+
