+++ src/main/resources/templates/registration.html
@@ -37,19 +37,16 @@
             <a class="navbar-brand header-title-big d-lg-block d-none">History in Books</a>
             <a class="navbar-brand header-title-little d-lg-none d-md-block">HIB</a>
             <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarCollapse"
-                    aria-controls="navbarCollapse" aria-expanded="false" aria-label="Toggle navigation">
+                    aria-controls="navbarCollapse" aria-expanded="false" aria-label="Toggle navigation" id="toggleBtn">
                 <span class="navbar-toggler-icon"></span>
             </button>
             <div class="collapse navbar-collapse" id="navbarCollapse">
                 <ul class="navbar-nav mr-auto">
                     <li class="nav-item active">
-                        <button class="btn pl-2 btn-header" id="link_main_header">Main</button>
+                        <button class="btn pl-2 btn-header" id="link_main_header" onclick="getHomePage()">Main</button>
                     </li>
                     <li class="nav-item">
                         <button class="btn pl-2 btn-header" id="link_books_header">Books</button>
-                    </li>
-                    <li class="nav-item" sec:authorize="hasRole('ADMIN')">
-                        <button class="btn pl-2 btn-header" id="link_admin_page">Admin</button>
                     </li>
                     <li class="nav-item">
                         <button class="btn pl-2 btn-header" id="menu-toggle">Category</button>
@@ -125,7 +122,7 @@
                             <input type="password" class="form-control" id="conformPassInputField"
                                    name="conformPassword" placeholder="Conform password" required>
                         </div>
-                        <button type="submit" id="hiddenSingUpButton" hidden>HiddenSingUpButton</button>
+                        <button type="submit" id="hiddenSingUpBtn" hidden>HiddenSingUpButton</button>
                     </form>
                 </div>
                 <div class="row px-4">
@@ -133,7 +130,7 @@
                         th:text="${errorMessage.getMessage()}" id="errorMessage"></h5>
                 </div>
                 <div class="modal-footer text-right">
-                    <button type="button" class="btn btn-primary col-12 mb-4 mx-0" id="singUpButton">Sign up</button>
+                    <button type="button" class="btn btn-primary col-12 mb-4 mx-0" id="singUpButton" onclick="sendSingUpForm()">Sign up</button>
                 </div>
             </div>
         </div>
@@ -159,29 +156,31 @@
 <div class="modal fade" id="signModal" tabindex="-1" role="dialog" aria-labelledby="exampleModalLabel"
      aria-hidden="true">
     <div class="modal-dialog" role="document">
-        <div class="modal-content modal-">
-            <div class="modal-header">
-                <h5 class="header-title-sign-in text-center" id="exampleModalLabel">Sign in</h5>
+        <div class="modal-content">
+            <div class="modal-header mx-3">
+                <h5 class="header-title-sign-in mb-0" id="exampleModalLabel">Sign in</h5>
                 <button type="button" class="close" data-dismiss="modal" aria-label="Close">
                     <span aria-hidden="true">&times;</span>
                 </button>
             </div>
             <div class="modal-body mx-lg-2 mx-sm-1">
-                <form>
+                <form action="/login" th:action="@{/login}" role="form" method="post" id="form-login" name="form-login">
                     <div class="form-group">
                         <label for="loginInput">Login</label>
-                        <input type="text" class="form-control" id="loginInput" placeholder="Login">
+                        <input type="text" class="form-control" id="loginInput" name="username" placeholder="Login"
+                               required>
                     </div>
                     <div class="form-group">
                         <label for="passwordInput">Password</label>
-                        <input type="password" class="form-control" id="passwordInput" placeholder="Password">
+                        <input type="password" class="form-control" id="passwordInput" name="password"
+                               placeholder="Password" required>
                     </div>
+                    <div class="custom-control custom-checkbox mb-2">
+                        <input type="checkbox" class="custom-control-input" id="checkboxRememberMe">
+                        <label class="custom-control-label" for="checkboxRememberMe">Remember me</label>
+                    </div>
+                    <button type="submit" hidden id="hidden_submit_btn">HiddenButtonToSubmitLoginForm</button>
                 </form>
-                <div class="custom-control custom-checkbox mb-2">
-                    <input type="checkbox" class="custom-control-input" id="checkboxRememberMe">
-                    <label class="custom-control-label" for="checkboxRememberMe">Remember me</label>
-                </div>
-
                 <div class="row">
                     <div class="col-lg-6 col-sm-12 pr-lg-1 mt-1">
                         <a class="btn btn-block btn-social btn-google text-white">
@@ -194,17 +193,17 @@
                         </a>
                     </div>
                 </div>
-                <div>
-                    <button class="btn btn-header border col-12 mt-3">Register new account</button>
-                </div>
             </div>
-            <div class="modal-footer">
+            <div class="modal-footer mx-3">
                 <button type="button" class="btn btn-secondary" data-dismiss="modal">Close</button>
-                <button type="button" class="btn btn-primary">Sign in</button>
+                <button type="button" class="btn btn-primary" id="sign_in_btn" onclick="sendSignInForm()">Sign in</button>
             </div>
+
         </div>
     </div>
 </div>
-<script src="../static/js/main.js"></script>
+
+<script src="../static/js/common-function.js"></script>
+<script src="../static/js/registration.js"></script>
 </body>
 </html>
